connectionhandler: tcp: unconditionally perform pal_close() on (re)init
The close_socket() was not calling pal_close() if the _running
was not on. This may be the reason for random failures on reconnect.
To be tested.

While here, move the socket_close() to be done before the error callbacks,
so the callback might be able to re-initialize the connection sequence.